add predictions plot

+++ erank/trainer/reptiletrainer.py
@@ -159,7 +159,8 @@ class ReptileTrainer(SubspaceBaseTrainer):
         # log epoch stats
         if self._log_train_step_every > 0 and epoch % self._log_train_step_every == 0:
             epoch_stats['meta-grad-norm'] = compute_grad_norm(self._model)
-            if self._log_additional_logs and epoch % (self._log_additional_train_step_every_multiplier * self._log_train_step_every) == 0:
+            if self._log_additional_logs and epoch % int(
+                    self._log_additional_train_step_every_multiplier * self._log_train_step_every) == 0:
                 if self._subspace_regularizer:
                     additional_logs = self._subspace_regularizer.get_additional_logs()
                     epoch_stats.update(additional_logs)
